Functions replaced in the component
